[declarative-transferengine] Add API and components to provide out-of-process sharing via sailfish-share. JB#53987 OMP#JOLLA-75
When allowing the user to share a file or some content, instead of
showing a SharingPage in-process, apps will now use ShareAction
to open the list of sharing options out-of-process.

The sharing options UI is launched in the sailfish-share app.
Transfer engine plugins now access sharing UI components from the
Sailfish.Share module instead of Sailfish.Transfer.