[declarative-transferengine] Allow opening and reading of file descriptors for ShareAction file resources. JB#54169 JOLLA-126
Add ShareAction::replaceFileResourcesWithFileDescriptors() to allow
file path resources to be converted to file descriptors, and
writeContentToFile() for the reverse conversion from file descriptors
to file paths.

This allows sailfish-share to open a file in-process and pass the file
descriptor to the sharing target app which may not have permission
to read the file path, but can read the file descriptor for the
purpose of sharing the data.