Implement 3rd party sharing
[share] Add QML API for 3rd party sharing. JB#56277 OMP#JOLLA-529

This plugin reads desktop files and builds SharingMethodInfo objects
that describe sharing to 3rd party applications. The applications may
expose a compatible D-Bus API for sharing using ShareProvider QML item.

Allow only plain text in ShareMethodItem. Less vulnerable.

Also there are error messages show in the UI if the D-Bus call fails.

When ShareProvider.triggered is emitted it gives a list of ShareResource
objects which can be in future expanded to support for example file
descriptor use or conversions of different types of resources.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>